Testing to see if we can match the names
A nested for loop that runs through the rookievorp.txt and the salaries.txt. It checks to see if the names match up. If yes, it says "He's in both", otherwise, it says "He's not in both". Eventually, we can mod this so that when the name matches up, we can put it into a new array, which can then be passed for the name parameter in the addPlayer function.