Update events.js
I added a team object for all the different positions and a function to add players to the teamObject. A loop is used to just see if players are being added. The output should be the last player added at each position along with their salaries. Also included is the salarytotal and vorptotal which come out as NaN for some reason.