Update events.js
This makes every name in both lists lower case so they are equivalent. This is done by using str.toLowerCase() on splitData[0] and splitData[1]. Also, The addPlayerToTeam() function now has an algorithm within each case. Each case initially adds the first player that matches the position. From there, any player that plays the position and has a higher Vorp will take over the position. The print outs show that the player with the highest Vorp was taken at every position except Right Field. This is because, there is still a discrepancy in one name for both lists. "Steven Souza" is the highest Right Fielder Vorp. However, his name in the salary list is "Steven Souza Jr." Assuming they are the same person, we need to find a way to delete or add on Jr.