Update events.js
pitcherAndCatcher object created. Each element of this object holds a pitcher, catcher, salaryTotal, and vorpTotal. This is every combination of pitcher and catcher possible. From here, we want to add the rest of the players. Then we should be able to filter out any salaries that are too big. Then finally, pick the team object with the highest vorp. This was worked on by Carl, Matt, and Mike. With help from Santiago as well.